+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Fix translate error in Error page](https://wealize.atlassian.net/browse/WLZAA-145)
+
 ## [v2021.04.27]
 
 - [Use a CDN]()